Added bound checking to keep the player on the sheet.
Also made the OpenGL and Geometry code for the PaperSheet static, since
they're all the same. This created a bug when onDestroy() is called, and
then onCreate(). Nothing ends up getting drawn. Regardless I need to
deal with onDestroy() in a future commit.

Also made some code in preparation for the option to choose difficulty.
Paper size will be one of the things affected by that.

	modified:   app/src/main/java/com/studioussoftware/paperescaper/game/GameManager.java
	modified:   app/src/main/java/com/studioussoftware/paperescaper/game/PaperGLRenderer.java
	modified:   app/src/main/java/com/studioussoftware/paperescaper/gameobjects/Camera.java
	modified:   app/src/main/java/com/studioussoftware/paperescaper/gameobjects/PaperSheet.java
	new file:   app/src/main/java/com/studioussoftware/paperescaper/model/BoundsChecker.java
	new file:   app/src/main/java/com/studioussoftware/paperescaper/model/Constants.java
	new file:   app/src/main/java/com/studioussoftware/paperescaper/model/Difficulty.java